Some fixes

+++ InstaView/SearchView.m
@@ -10,8 +10,7 @@
 
 @implementation SearchView
 
-static NSInteger const kSearchBarHeight = 44;
-
+static CGFloat const kSearchBarHeight = 44;
 
 - (instancetype)init
 {
